+++ rules/best-practices.js
@@ -48,6 +48,8 @@ module.exports = {
 		'no-fallthrough': 2,
 		// disallow the use of leading or trailing decimal points in numeric literals
 		'no-floating-decimal': 2,
+		// disallow assignments to native objects or read-only global variables
+		'no-global-assign': 2,
 		// disallow the type conversions with shorter notations
 		'no-implicit-coercion': 0,
 		// disallow var and named function declarations in the global scope
@@ -70,8 +72,6 @@ module.exports = {
 		'no-multi-spaces': 2,
 		// disallow use of multiline strings
 		'no-multi-str': 2,
-		// disallow reassignments of native objects
-		'no-native-reassign': 2,
 		// disallow use of new operator when not part of the assignment or comparison
 		'no-new': 2,
 		// disallow use of new operator for Function object