+++ rules/react.js
@@ -11,6 +11,8 @@ module.exports = {
 	'rules': {
 		// Prevent missing displayName in a React component definition
 		'react/display-name': 0,
+		// Forbid certain props on Components
+		'react/forbid-component-props': 2,
 		// Checks all JSX components and verifies that no forbidden propsTypes are used.
 		'react/forbid-prop-types': 0,
 		// Enforce boolean attributes notation in JSX
@@ -73,6 +75,8 @@ module.exports = {
 		'react/jsx-wrap-multilines': 2,
 		// Prevent usage of dangerous JSX properties
 		'react/no-danger': 0,
+		// Prevent problem with children and props.dangerouslySetInnerHTML
+		'no-danger-with-children': 2,
 		// Prevent usage of deprecated methods
 		'react/no-deprecated': 2,
 		// Prevent usage of setState in componentDidMount